Create README.md
This is the  linux implementation of the IICMPv4 Router Solicitation Protocol and ICMPv4 Router Adverisement Protocol
(RFC1256) using the C programming language. The ICMP Router Solicitation Message is sent by a computer host to any routers that may exist
on a local area network requesting them to advertise their presence on the network. This message is sent not only at bootup of the computer but also
periodically and is useful in that the computer administrator does not need to add IP routes manually to the host computer